[spack] move spack directory to tools
